display publish status
