+++ service/src/main/kotlin/fi/espoo/evaka/document/childdocument/ChildDocument.kt
@@ -99,15 +99,6 @@ data class ChildDocumentSummary(
     val publishedAt: HelsinkiDateTime?
 )
 
-data class ChildDocumentCitizenSummary(
-    val id: ChildDocumentId,
-    val status: DocumentStatus,
-    val type: DocumentType,
-    val templateName: String,
-    val publishedAt: HelsinkiDateTime,
-    val unread: Boolean
-)
-
 data class ChildBasics(
     val id: PersonId,
     val firstName: String,
@@ -120,6 +111,7 @@ data class ChildDocumentDetails(
     val status: DocumentStatus,
     val publishedAt: HelsinkiDateTime?,
     @Json val content: DocumentContent,
+    @Json val publishedContent: DocumentContent?,
     @Nested("child") val child: ChildBasics,
     @Nested("template") val template: DocumentTemplate
 )
